Fix Fremantle line not displaying in preview
In a980b64cea877d86da05f8b8efb426d3f286916d1, the station, Subiaco, was
changed because it was misspelled. However, the station is referenced
again, and had not be corrected. This meant the game was trying to link
Perth and Lock Street to a station that no longer existed.